+++ optalg/opt_solver/cplex_cmd.py
@@ -16,21 +16,21 @@ from .opt_solver_error import *
 from .opt_solver import OptSolver
 from .problem import OptProblem
 
-class OptSolverCbcCMD(OptSolver):
+class OptSolverCplexCMD(OptSolver):
 
     parameters = {'quiet' : False}
 
     def __init__(self):
         """
-        Mixed integer linear "branch and cut" solver from COIN-OR (via command-line interface, version 2.8.5).
+        CPLEX solver interface (via command-line interface).
         """
 
         # Check
-        if not utils.cmd_exists('cbc'):
-            raise ImportError('cbc cmd not available')
+        if not utils.cmd_exists('cplex'):
+            raise ImportError('cplex cmd not available')
         
         OptSolver.__init__(self)
-        self.parameters = OptSolverCbcCMD.parameters.copy()
+        self.parameters = OptSolverCplexCMD.parameters.copy()
 
     def supports_properties(self, properties):
 
@@ -45,21 +45,23 @@ class OptSolverCbcCMD(OptSolver):
 
     def read_solution(self, filename, problem):
 
-        f = open(filename, 'r')
-        
-        l = f.readline().split()
-        status = l[0]    
-        
+        import xml.etree.ElementTree as ET
+
         x = np.zeros(problem.c.size)
-        for l in f:
-            l = l.split()
-            i = int(l[0])
-            val = float(l[2])
-            x[i] = val
-        f.close()
+
+        tree = ET.parse(filename)
+        root = tree.getroot()
+
+        header = root.find('header')
+        status = header.get('solutionStatusString')
+
+        for var in root.find('variables'):
+            index = int(var.get('index'))
+            value = float(var.get('value'))
+            x[index] = value
+
         return status, x
 
-        
     def solve(self, problem):
 
         # Local vars
@@ -67,7 +69,7 @@ class OptSolverCbcCMD(OptSolver):
 
         # Parameters
         quiet = params['quiet']
-
+        
         # Problem
         try:
             self.problem = problem.to_mixintlin()
@@ -81,7 +83,7 @@ class OptSolverCbcCMD(OptSolver):
             input_filename = base_name+'.lp'
             output_filename = base_name+'.sol'
             self.problem.write_to_lp_file(input_filename)
-            cmd = ['cbc', input_filename, 'solve', 'solution', output_filename]
+            cmd = ['cplex', '-c', 'read', input_filename, 'optimize', 'write', output_filename, 'quit']
             if not quiet:
                 code = subprocess.call(cmd)
             else:
@@ -91,17 +93,17 @@ class OptSolverCbcCMD(OptSolver):
             assert(code == 0)
             status, self.x = self.read_solution(output_filename, self.problem)
         except Exception as e:
-            raise OptSolverError_CbcCMDCall(self)
+            raise OptSolverError_CplexCMDCall(self)
         finally:
             if os.path.isfile(input_filename):
                 os.remove(input_filename)
             if os.path.isfile(output_filename):
                 os.remove(output_filename)
+            if os.path.isfile('cplex.log'):
+                os.remove('cplex.log')
 
-        if status == 'Optimal':
+        if 'optimal' in status.lower():
             self.set_status(self.STATUS_SOLVED)
             self.set_error_msg('')
         else:
-            raise OptSolverError_CbcCMD(self)
-
-        
+            raise OptSolverError_CplexCMD(self)